refactor: ビジネスロジックの分離と設定変数のベンダーニュートラル化 (#37)
1. generate_missing_link_suggestions を core/llm.py から api/services.py に移動
   - llm.py はAdapter層の責務のみに限定
   - services.py からは generate_text() を呼ぶだけで動作

2. Config変数名をベンダーニュートラルに変更
   - openai_api_key → llm_api_key
   - openai_analysis_model → llm_analysis_model (デフォルト: o3-mini)
   - openai_embedding_model → llm_embedding_model
   - AliasChoices で旧 OPENAI_* 環境変数も後方互換で受け付ける
   - docker-compose.yml も LLM_* 優先、OPENAI_* フォールバックに対応
