+++ .env.example
@@ -1,6 +1,7 @@
-OPENAI_API_KEY=sk-your-api-key-here
-OPENAI_ANALYSIS_MODEL=gpt-4o
-OPENAI_EMBEDDING_MODEL=text-embedding-3-large
+# LLM 設定 (ベンダーニュートラル。後方互換のため OPENAI_* も使用可)
+LLM_API_KEY=sk-your-api-key-here
+LLM_ANALYSIS_MODEL=o3-mini
+LLM_EMBEDDING_MODEL=text-embedding-3-large
 JWT_SECRET=change-this-to-a-random-secret
 ADMIN_PASSWORD=your_secure_password
 DATABASE_URL=postgresql://episteme:episteme@postgres:5432/episteme
