+++ backend/api/services.py
@@ -840,3 +840,75 @@ def save_cb_session(
             session.close()
     except Exception:
         logger.exception("Failed to save course builder session %s", session_id)
+
+
+# ---------------------------------------------------------------------------
+# Missing Link Suggestion（分野横断パターン検索クエリ生成）
+# ---------------------------------------------------------------------------
+
+
+def generate_missing_link_suggestions(
+    pattern_name: str,
+    pattern_description: str,
+    structural_rules: list[str],
+    variables_template: list[str],
+    existing_fields: list[str] | None = None,
+) -> dict:
+    """パターンメタデータを受け取り、構造的空白を検知して分野横断の検索クエリを生成する。
+
+    Returns a dict matching the MissingLinkSuggestion schema (without pattern_id).
+    """
+    rules_text = "\n".join(f"  - {r}" for r in structural_rules) if structural_rules else "  (none)"
+    vars_text = ", ".join(variables_template) if variables_template else "(none)"
+    existing_text = ", ".join(existing_fields) if existing_fields else "none known"
+
+    prompt = f"""You are a cross-domain research advisor for the Episteme Graph system.
+
+Given the following abstraction pattern, suggest academic fields where this structural pattern
+likely occurs but is NOT yet represented in our pattern library.
+
+## Pattern Information
+- **Name**: {pattern_name}
+- **Description**: {pattern_description}
+- **Abstract Variables**: {vars_text}
+- **Structural Rules**:
+{rules_text}
+- **Fields already covered**: {existing_text}
+
+## Your Task
+1. Identify 3-5 academic fields/domains where this same structural pattern likely manifests,
+   but which are NOT in the "already covered" list.
+2. For each field, explain WHY this pattern would appear there (concrete reasoning, not generic).
+3. For each field, provide 2-4 arXiv search keywords that combine the pattern's structural
+   concepts with field-specific terminology. Keywords should be specific enough to find relevant
+   papers, mixing both generic structural terms and specialized domain terms.
+
+## Output Format (strict JSON)
+Return ONLY a JSON object with this structure:
+{{
+  "suggestions": [
+    {{
+      "field": "<academic field name>",
+      "reasoning": "<1-2 sentences explaining why this pattern appears in this field>",
+      "keywords": ["<keyword1>", "<keyword2>", "<keyword3>"]
+    }}
+  ]
+}}
+
+Important:
+- Do NOT include fields already covered.
+- Keywords must be suitable for arXiv search (English, technical terms).
+- Balance generic structural terms with field-specific jargon to mitigate hallucination."""
+
+    raw = generate_text(
+        messages=[{"role": "user", "content": prompt}],
+    )
+
+    # Strip markdown code fences if present
+    cleaned = raw.strip()
+    if cleaned.startswith("```"):
+        lines = cleaned.split("\n")
+        lines = [l for l in lines if not l.strip().startswith("```")]
+        cleaned = "\n".join(lines)
+
+    return json.loads(cleaned)
