+++ backend/core/llm.py
@@ -20,7 +20,6 @@ Reasoning モデル（o1, o3-mini, gpt-5.x 等）向けの自動変換を行う:
 
 from __future__ import annotations
 
-import json
 import logging
 import re
 from functools import lru_cache
@@ -37,8 +36,6 @@ T = TypeVar("T", bound=BaseModel)
 # ---------------------------------------------------------------------------
 # Reasoning モデル判定
 # ---------------------------------------------------------------------------
-
-_REASONING_MODEL_PREFIXES = ("o1", "o3", "o4")
 
 _REASONING_MODEL_PATTERNS = re.compile(
     r"^(o1|o3|o4|gpt-5)"
@@ -113,12 +110,12 @@ def _get_openai_client():
     from openai import OpenAI  # ベンダ依存を本モジュール内に封じ込める
 
     settings = get_settings()
-    if not settings.openai_api_key or settings.openai_api_key.startswith("sk-your"):
+    if not settings.llm_api_key or settings.llm_api_key.startswith("sk-your"):
         raise EnvironmentError(
-            "OPENAI_API_KEY が設定されていません。"
-            " .env ファイルに OPENAI_API_KEY=sk-... を追記してください。"
+            "LLM API キーが設定されていません。"
+            " .env ファイルに LLM_API_KEY=sk-... (または OPENAI_API_KEY=sk-...) を追記してください。"
         )
-    return OpenAI(api_key=settings.openai_api_key)
+    return OpenAI(api_key=settings.llm_api_key)
 
 
 # ---------------------------------------------------------------------------
@@ -152,7 +149,7 @@ def generate_text(
         LLM のレスポンステキスト。
     """
     settings = get_settings()
-    model_name = model or settings.openai_analysis_model
+    model_name = model or settings.llm_analysis_model
     client = _get_openai_client()
 
     adapted_messages = _adapt_messages_for_model(messages, model_name)
@@ -195,7 +192,7 @@ def generate_text_with_structured_output(
         パースされた Pydantic モデルインスタンス。
     """
     settings = get_settings()
-    model_name = model or settings.openai_analysis_model
+    model_name = model or settings.llm_analysis_model
     client = _get_openai_client()
 
     adapted_messages = _adapt_messages_for_model(messages, model_name)
@@ -235,79 +232,8 @@ def generate_embeddings(
         return []
 
     settings = get_settings()
-    model_name = model or settings.openai_embedding_model
+    model_name = model or settings.llm_embedding_model
     client = _get_openai_client()
 
     resp = client.embeddings.create(model=model_name, input=texts)
     return [e.embedding for e in resp.data]
-
-
-# ---------------------------------------------------------------------------
-# Missing Link Suggestion — LLM prompt engineering
-# ---------------------------------------------------------------------------
-
-def generate_missing_link_suggestions(
-    pattern_name: str,
-    pattern_description: str,
-    structural_rules: list[str],
-    variables_template: list[str],
-    existing_fields: list[str] | None = None,
-) -> dict:
-    """パターンメタデータを受け取り、構造的空白を検知して分野横断の検索クエリを生成する。
-
-    Returns a dict matching the MissingLinkSuggestion schema (without pattern_id).
-    """
-    rules_text = "\n".join(f"  - {r}" for r in structural_rules) if structural_rules else "  (none)"
-    vars_text = ", ".join(variables_template) if variables_template else "(none)"
-    existing_text = ", ".join(existing_fields) if existing_fields else "none known"
-
-    prompt = f"""You are a cross-domain research advisor for the Episteme Graph system.
-
-Given the following abstraction pattern, suggest academic fields where this structural pattern
-likely occurs but is NOT yet represented in our pattern library.
-
-## Pattern Information
-- **Name**: {pattern_name}
-- **Description**: {pattern_description}
-- **Abstract Variables**: {vars_text}
-- **Structural Rules**:
-{rules_text}
-- **Fields already covered**: {existing_text}
-
-## Your Task
-1. Identify 3-5 academic fields/domains where this same structural pattern likely manifests,
-   but which are NOT in the "already covered" list.
-2. For each field, explain WHY this pattern would appear there (concrete reasoning, not generic).
-3. For each field, provide 2-4 arXiv search keywords that combine the pattern's structural
-   concepts with field-specific terminology. Keywords should be specific enough to find relevant
-   papers, mixing both generic structural terms and specialized domain terms.
-
-## Output Format (strict JSON)
-Return ONLY a JSON object with this structure:
-{{
-  "suggestions": [
-    {{
-      "field": "<academic field name>",
-      "reasoning": "<1-2 sentences explaining why this pattern appears in this field>",
-      "keywords": ["<keyword1>", "<keyword2>", "<keyword3>"]
-    }}
-  ]
-}}
-
-Important:
-- Do NOT include fields already covered.
-- Keywords must be suitable for arXiv search (English, technical terms).
-- Balance generic structural terms with field-specific jargon to mitigate hallucination."""
-
-    raw = generate_text(
-        messages=[{"role": "user", "content": prompt}],
-    )
-
-    # Strip markdown code fences if present
-    cleaned = raw.strip()
-    if cleaned.startswith("```"):
-        lines = cleaned.split("\n")
-        lines = [l for l in lines if not l.strip().startswith("```")]
-        cleaned = "\n".join(lines)
-
-    return json.loads(cleaned)
