+++ docker-compose.yml
@@ -58,9 +58,9 @@ services:
       - "8001:8001"
     environment:
       JWT_SECRET: ${JWT_SECRET:-episteme-dev-secret-change-in-prod}
-      OPENAI_API_KEY: ${OPENAI_API_KEY}
-      OPENAI_ANALYSIS_MODEL: ${OPENAI_ANALYSIS_MODEL:-gpt-4o}
-      OPENAI_EMBEDDING_MODEL: ${OPENAI_EMBEDDING_MODEL:-text-embedding-3-large}
+      LLM_API_KEY: ${LLM_API_KEY:-${OPENAI_API_KEY}}
+      LLM_ANALYSIS_MODEL: ${LLM_ANALYSIS_MODEL:-${OPENAI_ANALYSIS_MODEL:-o3-mini}}
+      LLM_EMBEDDING_MODEL: ${LLM_EMBEDDING_MODEL:-${OPENAI_EMBEDDING_MODEL:-text-embedding-3-large}}
       NEO4J_URI: bolt://neo4j:7687
       NEO4J_AUTH: neo4j/episteme
       DATABASE_URL: postgresql://episteme:episteme@postgres:5432/episteme
